Clear out table model before displaying new results
After a  GUI based judged run completes, it fills in the JTable data model with the new results asynchronously, but the table is shown before this completes.  As a result, stale data could be seen from the previous judge run.  This change clears out the table model and replaces it with a single line indicating that "Judging" is still taking place (in actuality, the table model is being populated with the new results, and that takes a relatively long time).  Replacing the old stale data with a single line telling the user what is going on is more user friendly and certainly less misleading.