Create separate methods to fetch API Version
The Test Shadow button, in addition to making a simple "connect" test, will also attempt to query the version information of the API from the remote.  The code to do this was originally added in-line and made the code difficult to read.  It has now been broken into separate methods to break the operation of getting the version info into distinct pieces for readable, and possibly unit testing someday.