Fix scrolling and viewing of new entries
The pane that contains the connection status JTable would not previously show the newest entries as they were added.  The user had to manually scroll down to see them.  Now, the pane will move the viewport to show to bottom most entries in the table as the content in the pane changes size (gets bigger when more entries are added).