Add Status column to shadow status table
At @lane55's suggestion, in an effort to make the status JTable easier to read, a column was added that simply indicates the color-coded status of the notification messages: SUCCESS, FAILURE or INFO.  As a result of doing this, the actual messages displayed were changed so as not to duplicate the words failure or success.
To further increase readability of the status table, it has been changed so it will get bigger if the user decides to resize the window.  This will allow you to view more rows at once.  This involved changing the layout to GridBagLayout instead of FlowLayout in the "center" pane.