Move retry test and delay logic to main connect loop
The shadow connect loop did not always check if connection attempts were desired, and consequently, always attempted to reconnect with no delay.  This was wrong.  In all cases of a disconnect or failed connection, whether or not to retry should be checked, and then, if retries are desired, the sleep should be done.  Previously, this was ONLY done if a connection succeeded at one point and THEN failed.