Add back old (now deprecated) constructor to satisfy unit tests
The Unit tests extend the Executable class and then proceed to call the base class constructor super(...) directly.  The problem is, that constructor was removed when the execution frame was added.  For the purposes of the junit tests, which do not use the frame (or gui), the original constructor was added back, and it sets the execute frame to null and use gui to false (which it should be already).