Add support for entry_point
Make use of previously added "entryPoint" member variable in the Run class.
Involved adding new submit method to IInternalController that takes the entry point (a string), or null if no entry point.
Change implementations of all classes that implement IInternalController.
Update Event Feed submission logic to pass the entry_point if there is one.
Make sure to replace the :basename with the entry point prior to substituting variables.
The already SubmitterFromEF class may be able to be used instead of the already existing RemoteRunSubmitter class.  Apparently the SubmitterFromEF class was made to allow command line submissions for a command line tool using the EF.  Should be refactored, probably.