Add conditional suffix substitution variable and test unit
Add the {:ifsuffix=xxxx} substitute variable allowing conditional substitution if the existing string suffix at the location of the variable is not the same.  Useful for Kotlin (etc) to add Kt to {:basename} if not there.  From the event-feed, the entry_point specifies Kt on the main class.  For running from GUI or AJ, it must be added if not there.
Add test unit to exercise new substitute variable.