Fix case where compiler output file name does not match
For some languages (Kotlin, for example), the {:basename} is not the name of the "executable" output of the compiler.  Rather, it is mangled - possible case change, and appending of "Kt".  This fix allows the use of the entry point as well as the conditional suffix addition for compiler output.  This allows you to uncheck the "interpreted" check box for Kotlin and Java with packages and a specified entry point.