CI: Add Entry Point tab to View Source
If an entry point was specified for a run, there was no way to see what it was.  This fix adds an "Entry Point" tab to the "View Source" MultipleFileViewer that shows the entry point, if one was specified.  The tab is not there if no specific entry point was specified.