Fix the terminate/stop auto judging button flashes
For the AutoJudgeStatusWindow, the Stop Autojudging/Terminate button will not flash back and forth for each test case.  Rather, the Terminate button will appear when Execution starts and will not revert back to Stop Auto Judging until after the last test case has been validated.  This involved changing when the button state was set. Previously it was switched back and forth between test cases.
Changed executionFrame member to be executionMonitor since that is what it is now that the interface has changed name.