Fix deficiencies found while testing for NAC23
The service install script and sandbox setup scripts have to set the owner of the cgroup to the user/group that will be doing judging.  As such, the service install script now takes 2 arguments specifying this user and group.  Previously it was hardcoded as pc2/pc2.  For NAC, however, we ubuntu/ubuntu.  So, we needed a way to specify that at system setup time.