Fixed execute / validate message length changing
Change status font to fixed width in ExecuteTimerFrame and AutoJudgeStatusFrame.
Make sure length of string passed as run status is fixed length.  Created formatTestCasePhase() method to perform constant width formatting suitable for up to 99999 test cases.
Note that this is purely for aesthetics; if the string is not constant length, the message label will simply show text changing in length which can be irritating.