Better error reporting and using correct install directory
With the necessity to specify a user and group for cgroup ownership, arguments were added to the script but not properly documented.  The usage message is now much clearer.  There is also a detailed comment in the script indicating what the arguments mean.
The script used to believe that /usr/pc2 was the only place PC2 could be installed.  This has been changed to figure out where pc2 is installed and use that folder instead of /usr/pc2.  pc2env is invoked to obtain $basedir
Turn off debugging, which was on by default.