Handle sandbox default for adding a new problem
When a problem is added, a null problem is passed into EditProblemSandboxPane.  This previously caused an NPE.  Now, if the problem is null, we try to set up the default values for the sandbox based on the constants in the Problem class.
Also, add an exception handler and log the exception and display an error box if an add problem should fail in the future, as opposed to silently ignoring the problem.