CI: correctly copy and convert bytes to Kbytes in problem table
When a problem is copied to create a new problem, the max output size was not copied.  As a result, 0 was set as the limit.  While this is an error, it set the limit 0, which signifies the system default limit.  However, the problem table incorrectly displayed this value as bytes, not KBytes. This resulted in a displayed value that was 1024 times bigger than it should have been.
When a problem is copied, copy the max output size in KB to the new copy.