Make unsupported sandbox nag message user controllable
If sandboxes are not supported, a nag message informing the user of this is presented every time a problem is edited.  A checkbox was added to allow the user to not show the message for future problem edits (in the current instance of the application).
In addition, superfluous setProblem() calls to the sandbox edit pane were removed because the sandbox edit pane was getting populated twice for every problem edit.
 The sandbox edit pane has the problem information set in the EditPRoblemPane.populateGUI() which is called every time that EditProblemPane.setProblem() is called, so there is no need to set it elsewhere.