Correctly handle missing teams.tsv if teams2.tsv is present
If teams2.tsv was present, but teams.tsv was not, pc2 would not load the CDP, indicating that teams.tsv was missing.  teams.tsv is NOT required if teams2.tsv is present.
Added comment to checkFiles() routine to describe what the routine does.
Removed extra / duplicate code that was already done below, and, shouldn't have been there in the first place.