Refactor newly added enum
The JudgeFileType enum already exists in the Utilities class by a different name, so use that.
Refactor code to make use of new getJudgeFileName() method.
If an error occurs when looking up a judge's file, show the expected name of the file, if it's available.  This will aid debugging.
Amend javadoc comment in Utilities.locateJudgesDataFile() to indicate that the last argument is never used, since the serialized file is already passed in.