CI: Check for bad run state when starting up
During server startup, if a run has a status of JUDGED and there are no JudgementRecords, set the status to NEW because NPE's will result otherwise.
This can happen if loading an older contest where there were no restrictions on what the run status could be changed to using the Edit Run pane.