Add legacy fromtSVFile() method for unit tests
Unit tests do not have an IInternalContest object so, we provide a shim method to pass a null contest object to the real from TSVFile(IInternalContest,...) method.  This works fine for existing Unit tests, however, if the units want to test out setting/changing institution codes, then a real contest with a real CDP path will need to be supplied.