Remove unnecessary restriction on properties in contest PAUSE message
A non-compliant restriction requiring only 2 properties in the contest start time pause json was causing 405 errors to be reported to remote clients indicating the presence of extra data in the request.  Specifically, the contest_pause_time property was present.  Even if PC2 does not use the property, it should be allowed - extra properties should not preclude the success of the call.