CI: Fix incorect separator for section.key
There are 2 getValue() methods.  One takes a single string already formatted Section.key to look up.  The other takes 2 strings: the section and the key and is supposed to put a period between them.  Rather, it was putting a colon between them. The second method is never called in any of the PC2 projects, but it should be fixed (or removed?)