Validate that memory limit is non-negative
Note that this test is ALWAYS performed, even if a sandbox is not selected for the problem.  It is possible that the user selected a sandbox, entered a negative value, then de-selected a sandbox.  We STILL want to issue an error here so we don't have any future issues with the memory limit.  Note that this scenario would require the user to perform the above sequence, so they went out of their way to put a negative value there, which we simply will not permit.