i843 Add support for group filtering in scoring algorithms
In addition to the already present Division Number scoreboard filtering, also allow a Group to be specified.  This allows filtering based on division number, group number, group and division number of no filtering.
Added group filtering to support CLICS API 2023-06 (and 2022-07).  This commit only includes the changes to the scoring algorithms.