i843 deal with optional QueryParams for scoreboard API endpoint
The CLICS 2023-06 spec (and 2022-07) allows you to specify a query param of ?group_id=XXXX to give a scoreboard for only the specified group id.  PC2 also supports the notion of divisions where several different groups may belong to a single division.  To support this, an additional optional query param "?division=#" was added.  Specifying the division query param will return a scoreboard for all teams in groups that belong to given division number.