i871: update AuthService.isLoggedIn() to also check for "current page".
The point of this was to have "isLoggedIn()" be based not only on the
presence of a "token" (which comes back from the WTI-API server upon
login) but now also on the existence of a "currentPage" in
sessionStorage, which would indicate that this was a restart due to
(e.g.) a user hitting "F5".

However, it's not clear that simply checking for a currentPage is
sufficient; does the old "token" get lost on restart?  Is the old
websocket still open?  These questions still need to be resolved.