i843 add awards endpoint handler
CLICS 2023-06 Awards endpoint is quite involved.  It allows anyone to request the awards, but, it must obey freeze time.
This involved changing existing routines (CLICSAwardUtilities and ScoreboardUtilities) that compute the awards (for the awards comparision) to accept a boolean indicating whether or not to obey freeze time.
Had to add methods to accommodate previous usage that assumed obey freeze time was false (for awards comparison and test units).
Fixed unit test so pass false as obey freeze time.