i888 Use correct java compiler for JUnit tests
Since Eclipse 4.17 (2020-09), Java 11 is the default JDK that ships with Eclipse.  Previously, it was Java 8 (1.8).
This PR attempts to track down the correct JDK to use for compiling the JUnit test programs.  The JDK must match the currently running JVM in order to qualify.  A search of  the JAVA_HOME/bin and folders on the PATH environment variable are checked for the presence of the corresponding JDK (javac[.exe]).
The code only checks for the presence of the java version string in the folder name.  It could be expanded to actually attempt to run "javac -version" in each folder, but that shouldn't be necessary.