i843 Add additional submitClarification method specifying clientId
When clarification submissions come in over the API, it is necessary to specify the clientId of the submitter (not the user logged into the model, which would be a feeder account at the moment).  In addition, an administrator or proxy user can submit a clarification as a team, so there had to be a way to create a clarification submission as a specific client.
Just added another method to the controller interface that takes a ClientID of the submitter.
Updated all classes that implement IInternalController to implement this new method.