i900 Add notion of a primary group id
The CMS currently allows only 1 group (site) to be defined per use.  We remember which group was the one provided by the CMS.  This is used for display format substitution for team names that want to include the group.
TODO: When the CMS supports multiple groups (sites) per user, the display format substitution routine will have to be modified, or, the ability to add a group to the display name will have to be removed.