i900 -General code cleanup
Fixed some copyright comments.
Removed extra spaces at ends of lines
Removed dunsel code
Consolidated some scoring methods to ScoreboardUtilities

Note: in DSA and NSA, groupRank is only computed for the primary group (CMS group).  This should be changed to compute the rank for each group the team is a member and save it in the standingsRecord.  Then, in the XML create a groupinfo element for each group the team is in contain it's group rank for that group.  See addGroupRow in NSA.