i_900 Add group filtering to support filtered event feed
Add group filtering to the Filter class.  This allows reports to be filtered by group.  Particularly useful on the event feed json report (for resolver).
Note that if a Filter has group filtering enabled, this will apply group filtering to accounts and problems, even if the filter is not explicitly filtering by account or problem.  This is logical.