i900 Division filtering bug fixes
There was a longstanding bug where if a team was not a member of a group that ended in D#, the scoreboard (DSA) would cause a NPE and never generate a scoreboard.  This was because a null is returned as the 'div' if the group (or now groups) the team is in is not a "Division" group (*D1, *D2, ...)
Fixed bug with wrong variable check in loop for groupname.