i_900 Include groups of filtered teams in EventFeed
Previously, only the since group in the filter for the event feed was included in the event feed "groups".  In reality, any groups that the filtered teams belong to should also be included in the event feed.  Consider a team that is a member of groups D1, Oregon, Oregon-D1 and we filter the Eventfeed by group D1.  There were no entries in the eventfeed for Oregon nor Oregon-D1, even thought the team was a member of these.   This fix includes those "groups" entries in the eventfeed.