i933 Explicitly check for server when initializing for RunsTablePane/RunsTable
Add an explicit check for a client of type SERVER when setting the contest and controller in SelectJudgementFrame.  This Frame is used by RunsTablePane and RunsPane.
We can not simply use isServer() since 1) SelectJudgementFrame is not a JPanePlugin and 2) any panes that ARE created are created later, so they are not available yet.
The test is simple enough to determine if the client is a server or not.