i_815 Read-only feeder clients should not attempt to start/stop the contest
When a state "started" or "ended"events are received, read only feeder clients (feeder2, feeder3, ...) would attempt to start (or stop) the contest.  This was fixed by using the already existing isReadOnlyClient() method to only allow non read-only client(s) to start and stop the contest.  The way it was, feeder2, feeder3 would get pop-up error messages indicating they did not have permission to start the contest.  But, if they did have the permission, it could cause issues, so, just prevent read-only clients for starting or stopping the contest due to state events received.