i_815 Fix issue with tossing judgments after contest ends
When the code was added to allow shadowing to start before the contest starts, code was also added to prevent both submissions and judgments from being handled if the contest isn't running.  This has to be changed to separate the criteria for submissions and judgments, since judgments may come in after the contest has ended.  Added a couple of shorthand variables to test for the contest having been started previously (wasContestStarted) and if the contest is currently running (isContestRunning)
CI: shorten messages that appear in the GUI status table