i_942 Fix entry points with periods
If an entry_point was specified in a "submissions" event on the event-feed that contained a package (for Java) or an explicit file (for Python3), the run would fail (compile error for Java and RTE for Python3).
Add code to handle the special cases for each language that can specify an entry point.
Code works the same as before if no entry_point specified.
Remove pruning of entry_point when it is set in the Run.   This was removing important information for Java entry oints with package information specified and extensions on python files.
CI: Add constant for {files} command substitution string and change in places it's used.