i943 Fix Groups tab behaviour on Edit Problem dialog
Do not enable/disable individual group check boxes.  This causes excesseive overhad when checking for the Update button.
Instead, enable/disable the entire JCheckboxJList control which CONTAINS the checkboxes.  This appears to be the original author's intent.
Left some commented-out code with comments to help in code review.  Will be removed prior to merging after approved.