i_232 Finishing touches
Add ability to have all judges' samples "not stop on failure" by default, no matter how  a problem has defined the setting.  This can be controlled by a checkbox on the SampleResultsPane.
Add the one-shot overrideStopOnFailure flag for a Run and handle it in Executable
Add submitJudgeRun methods to support the overrideStopOnFailure flag to InternalController
Add new method to PacketFactory to pass the overrideStopOnFailure flag to the properties for a run submission.
Fix PacketHandler to deal with the overrideStopOnFailure property and set it in the Run so Executable will see it.
Bug fix for filter by judging tyeps
CI: Add ability to show why a run failed in the TestResutlsPane instead of just showing "Fail"