i_232 Fixes for Junit tests and indenting
Changed <tabs> that were in original code to spaces to fix indenting.
Fixed bug in packethandler when overrideStopOnFailure is not present i the properties.
Cosmetic/spacing changes added by Eclipse + @overrides
Changes to support unit tests.
Re-add legacy language support to LanguageUtilities.  This should be deprecated, but there may be users out there that do not supply CLICS ID's in the language definition yaml, so for now we have to include it.  This was found with unit testing.