i_972 Add support for better human judge reviewing
Modify scripts to generate information to help a human judge figure out why a run failed.
Better determination of which CPU to run a submission on in a sandbox.
Add some basic Web page scripts to make judge results web pages on each judge machine.
Add a configurable execute folder that allows substitute variables.  This is good so that the execute folders may be retained for each run {:runnumber} in the execute folder, for example.
CI: toString method of ExecutionData was not correct.