i_966 fix results.tsv and scoreboard.json comparisons
Use natural sort compare for json/tsv fields, so x1 appears before x[10]
Add compare of problem information for scoreboard.
Add constants to control if certain conditions will cause a comparison difference (addresses CDS issue and PC2 issue)
For now, clean up problem id field by converting it back to a string from an elementid.
Do not abort entire compare process if a file is missing.  note the missing file in the output as a failure, but continue doing the other files.
CI: add line number if errors occur during TSV processing (this was an omission in original code)
Change order of presentation of failures in the report file to coincide with the number of failures printed in the beginning.
Add natural compare routine to StringUtilities.