i871: use "IWebsocketService" instead of "WebsocketService". Result:
this forces the use of the WebsocketServiceFactory to generate a
WebsocketService object.  Otherwise, TypeScript "constructor parameter
properties" will implicitly perform a manual construction of a
WebsocketService, which (a) bypasses the ability to use the Mock
service, and (b) causes there to be TWO "websocket service" objects
created -- one from the factory, the other from the implicit constructor
parameter construction of a local "property" variable.  These two
objects cause conflicts when the Angular SPA state is restored following
an F5.