i1027, i1029: update "contest is running" req for viewing clars & probs:
This commit removes the requirement that the contest be running in order
for the WTI-API /clarifications endpoint to return clars; now the only
requirement for returning clars is that the client "is-logged-in".

The commit also changes the requirement in the /problems endpoint:
previously the contest had to be RUNNING in order to see problems; now,
it just has to have been STARTED (i.e., there must have been greater
than zero "elapsed time".  This allows teams to see problems during a
pause.

Note that this commit is part of a PR for
https://github.com/pc2ccs/pc2v9/issues/1027, but it solves
https://github.com/pc2ccs/pc2v9/issues/1029 as well.