i1030: set announcement clar state for empty OR "announce" question:
previously, announcement clars were only set to state "ANNOUNCED" if the
question was empty.  Empty questions are confusing on the team grid, so
announcements were updated to contain a "question text" value of
"Announcement".  This commit changes the setAnswer() methods to set the
state to "ANNOUNCED" if the Announcement Question text is EITHER empty
OR "announcement" (ignoring case).