i1030: replace combobox ItemListener with ActionListener, because...
...an "ItemListener" is invoked twice when a combobox changes (once for
the unselected item; again for the newly-selected item) whereas we're
only interested in the fact that an action selected a new item.

Also removed a redundant adding of groupsAndTeamsList to the groupsPanel
(it was being added twice but should only be added once).
Also cleaned up some documentation and layout sizes, and refactored some
variable and method names.

Finally, added a "TODO" comment on one place where the code doesn't make
any sense and it potentially VERY inefficient.