i1030: minor additions to SubmiClarPane.  However:
this class should probably be reverted all the way back to its original
version, before any modifications were made to it for supporting
"announcements".  A completely new class, "GenerateAnnouncement", has
been implemented and added to the JudgeView pane to support
announcements; this class has been removed from the JudgeView -- it
should be added (in its original form) to the "Developer's Pane".