i_843 i_1048 Allow full CLICS RELTIME values for contest.yaml durations
The YAML parser will convert any value it sees with a fractional part (decimal point followed by numbers) to a double value, which is converted to a String in ContestSnakeYAMLLoader.  The causes a problem when that value was parsed as a Long (it caused a NFE).  This fix converts the value to a double using (Double.parseDouble()), then, it typecasts that to a long.  This is OK, since contest durations are in seconds, not fractional parts of  second.  We just ignore the fractional part.
Also fixed JUnit test correctly since the freeze duration will now compare properly without stripping off the leading 0.
This fix will be part of the PR for i843_CLICS_API_Endpoint_Compliance, but can be cherry-picked elsewhere if needed before this is merged.