i_843 CI Add pc2Site property to ScoringGroup
There was a bug that occurred during standing calculation if a Group was assigned a site when it was created.  In general, we don't assign sites to groups, but if you did, it would cause an unknown property (pc2Site) during the XML deserialization.  Adding the benign property to the ScoringGroup class addresses this issue.