i_843 - Trigger a state change on contest entering frozen state
To correctly implement the CLICS event feed, a contest "state" notification must be sent when a contest transitions to the "frozen" state.  This code was incorporated into the refactored ContestClockMonitorThread (formerly AutoStopContestClockThread).  In addition to checking for the end of the contest, the thread now checks if the contest enters the frozen state.