i_843 Only accounts with VIEW_EVENT_FEED can see CLICS API
Only add accounts to the Jetty Basic Authentication user table if the account is in realm.properties OR it's a PC2 account and the Permission.Type.VIEW_EVENT_FEED is enabled.
Add an AccountListener to update Jetty's Basic Auth account table if PC2 accounts are removed, added, permission changed or any other modification to the account.