i_857 Implement countdown pause time patch
Fix CLICS 202306 contest service to correctly handle the countdown_pause_time property.  This will simply clear the schedule start time (setting it to undefined), and remember when the contest was paused (in ContestInformation).
Fix contest clock displays to show the countdown is paused if it is.  Tooltip will show the time the contest was paused.