i_1058 Extra submission notification on event feed
Do not send an added submission notification if a run changes, only if the run is added.  If a run changes, it implies that was already added at some point, so why add it again.  Appropriate log messages are now provided in the feeder log when runs are added, and, in the case of a change where a run WOULD have been added.