Use supplied team_id for submissions by admins via the CLICS API
Use the optionally supplied team_id to make a submission on behalf of a team by an adminsitrator.  Previously, the team_id was validated, but never used.  We now create a ClientId and use that when we make the submission.
CI: expand error logging to include both the user making the request AND the team_id of the submission.
CI: expand error logging if a team is trying to submit on behalf of another team, or otherwise causes a permission error.