i_1065 dont send to proxy submitters twice
If a run is submitted by a proxy (not a team), then the proxy client who submitted it will receive two RUN_SUBMISSION_CONFIRM notifications.  One is sent directory using sendToClient() and the other is sent as a result of sendToJudgesAndOthers().  Test for this condition and prevent that from happening.