i_1071 fix account type on event feed
The CLICS 2023-06 specification says that administrator accounts should have type "admin" on the event feed and the API.  We previously just used the ClientID type, which was "administrator".  This fixes that.
Also, we manually set the other types so we do not depend on what PC2 uses for the client type string.