i_1080 Use Finalize tab settings for awards endpoint
If a contest is not yet finalized, the CLICS Award endpoint was always using WF awards 4g/4s/4b + WF ranking, even if different values were set on the Finalize tab in the administrator.
The awards endpoint should always use what settings may have been set on the Finalize tab, even IF the contest is not yet finalized.
Add a TODO that suggests creating a method that generates a default FinalizeData object that everyone can use a-la ResutlsFile.GenDefaultFinalizeData().