i_1080 Fix broken JUnit awards test
When the change for using FinalizeData was added, a bug surfaced in the JUnit that tests the awards generation.  One test created a custom FinalizeData (4,4,13).  It turns out these values were never previously used since the contest was not certified, and the defaults of 4,4,4 were used.  If the 13 had actually been used, the JUnit would have failed because the expected value was erroneously set to 11.  Previously, the custom FinalizeData was only used if the results were certified.  The results in the JUnit were never certified, so the custom values were never used.  When this PR was added, the values are now used if they are set (and the test for the contest being certified was removed).  As a result, the JUnit failed because the bronze count of 13 was being used.   I think the intent here was to use 13 as the expected value.
Also, the computation of ranks was wrong.   But, since they weren't being used anyway (previously) it didn't matter.  Now they ARE being used, so they had to be fixed.