i_764 Changes for submitting via shadow and CLICS API
The submitRuns() methods in InternalController/IInternalController had to be modified to use the throttling mechanism.  In addition, they now throw exceptions which must be indicated.
Fix message returned to a team submitting too fast using the API.  It was spaced improperly and it had internal PC2 java class information which is not needed by the user.  All the information is logged however.