i_841 Deal with kotlin lower case source file names
When kotlin compiles a file that starts with a lower case letter, it produces a class file that starts with an upper case letter.  As a result, the PC2 routine to check if the compile worked failed because, for example, testprogKt.class != TestProgKt.class.   These two changes first check the compile result filename for this condition, and then use the correct name when executing the result.