i_1114 Put correct version of pc2sandbox_interactive.sh in repo
Update repo with correct version of the interactive sandbox script.  A test was missing to kill a submission if the interactive validator finishes first.  Without this fix, we would sometimes notice JE (judging errors) or other judgment mismatches when comparing with a primary or shadow.  This fixes the so-called "DOMjudge" mode of interactive problem validation.