i_1116 add flags for Bill rules for standings entries
Add FTS to IProblemDetails details as a property.
Update reports and JUnits to print out FTS
Change nomenclature for medal places as opposed to ranks.
In DSA, remember if we excluded groups in the standings (that is, someone only wants standings for one or a couple of groups).  In this case we won't generate the Bill rules or medals.
Need to remember the XML memento for each teamStanding wo we can fill in the medals and Bill rules after the whole thing is generated.
Clean up new method in ResultsFile to calculate the ranks for Bill Honor rules.
CI change string compare with constant to avoid possibility of NPE