i_1116 Fixed ResultsFile bug  found by JUnit
The new method, createCitationRankInformation() in ResultsFile always expected at least as many teams in the standings as there were medals being awarded.  The JUnits pointed out this is necessarily the case.  As such, a safeguard was added to prevent calculation of WF Honors citations if there are less teams than medals in the standings.