i_1122 Create a catalog file during execution
The src_catalog.json file contains details about the contestant's source code, such as the original full path of the submission, whether the file is stored externally, etc.
This is useful for generating specific pages for the judge's submissions.  We can look at the full path of the submission source and see if it's from the CDP; if so, then we know they're judge's sample submissions.