i_971 Add options to not count CE/SV submissions
Other CCS's do not count CE (if its penalty is 0) or SV (if its penalty is 0).  PC2 always counted these as submissions.
There are now two new options in the ContestInformationSettings for controlling whether or not to count CE or SV as submissions (if their respective penalties are 0).
Changes to DSA and NSA to support this.
CI: fix bug in older JSON Standings report where it didn't put a comma between array records (it NEVER worked).