i_1109 Allow exception of bad values in scoreboard deserialization
Previously, Utilities.nullSafeToInt() and StringUtilities.getIntegerValue() were used to obtain the values of total_time and time.  Basically, those routines would assign a value of zero if the string was empty or null, and, in the former case, if a bad string to integer conversion occurred.  Now, since these are for score values, we will call Integer.parseInt() which will throw an exception if bad values are found.  All values should be good at this point and this shouldn't happen, Tammy.