i_1006 Adjust JUnits for new TestDataGroups
The JUnit tests for the ContestSnakeYamlLoader made some assumptions about the locations of the secret data.  This is now handled automatically by the TestDataGroup class.  Fix the routine in ContestSnakeYamlLoader to not put the "secret" folder on the directory path for data.
Fix spelling error in JUnit print message.
Make sure the static log is created before running junits since the TestDataGroup stuff will log things.