i_1006 allow LegacyGrader to accept list of judgments
If using the LegacyGrader internally (and not as an external process), the "String gradeTestCases(ArrayList<String> testCaseResults)" may be used directly.  The bypasses the extra steps of saving the list of judgment/score tuples to a file and having the grader open and read that file.  In addition, the gradTestCases method returns the net judgment string, eg. "AC 50" on success, and null on failure.

(cherry picked from commit b50c9610d21d1dbda6b99c19cd8d4115084c21de)