i1006: update Executable to support point-scoring.  Changes include:
- updating method executeAndValidateDataSet(testCaseNum) so that if the
contest is "point-scoring" then the test case score, judgement (acronym)
and TestDataGroup are stored in the RunTestCase (result) before storing
the RunTestCase in the Run.

- determining a "score" for each test case by looking for a "score.txt"
file left by the validator and using the result from that file if
present, otherwise using appropriate TestDataGroup rangeMax() or
rangeMin() value as the test case score.

- updating method execute(boolean) so that after all test cases have
been executed, if the contest is "point scoring" it invokes the
"LegacyGrader" to recursively determine the overall "score" and
"judgement" for the Run and inserts these into a new JudgementRecord in
the Run.