i_1006 Add ignore_sample option to LegacyGrader
Specification illudes to how to implement ignore_sample.  It relies on the fact that it may be used only at the root level and there must be 2 groups (sample and secret) at the root level. *sigh*  What a hacky way to implement it, but - that's the spec.
Add JUnits to test the flag - 3 cases.