i_1006 use test data groups for validateCDP
validCDP used to form the paths to judge's data on the assumption thta everything was in either sample or secret.  With test data groups, that is no longer the case, so change validateCDP to use the files from the ProblemDataFiles, which, is probably what it should have done in the first place.

(cherry picked from commit e3e8ae80f8482396b4eeddbe1fc682bf493db301)