i_1006 use correct test data group for max_score
We can't just use the first test data group and assume it's the root.  We have to walk up the getParent() chain until we find the null parent.  At that point, we know we're at the top of the tree, and we can use that getRangeMax().
As per Tim deBoer, he suggested we do not send Double.POSITIVE_INFINITY (infinity) if the max is in fact infinity.  Rather, just leave the property out of the record.  This contradicts the current 2023-06 specification that requires the max_score property.