i1006: save judgement acronym in ExecutionData (not acronym string):
the meaning of "acronym string" is ambiguous:  does it refer to the
string representation of the acronym element (for example "AC" for the
'AC' element of the acronym enum), or does it refer to the "description
string" associated with the acronym (for example, "Accepted" for 'AC').

Note that the enum methods toString() and name() return the former,
while the getValue() method returns a field (confusingly called "name")
which contains the latter.  To avoid confusion, only the acronym itself
is stored; clients can use getJudgementAcronym() to get the actual
acronym (enum element) and then use toString(), name(), or getValue()
depending on what they actually want.