i_1155 make sure judgment text matches configured types
Judgment result strings (judgment text) should only include those strings defined in reject.ini, or, the defaults if there is no reject.ini.  Some CLICS validators return "Wrong Answer" as a judgment string when the reject.ini specifically sets WA to be "wrong answer".  This causes confusion for contestants who are told they will only see a very specific set of judgment types.
Fixed bug where an MLE (memory limit exceeded) would be returned, even it it was not configured.  This is now mapped to RTE (run-time error) if MLE is not configured.
Make judge thick client operate the same as the AJ Monitor when deciding on the result string.  That is, it does a case insensitive compare now, as opposed to a an exact compare when testing result strings against those configured.