i_1179 Send encoded ZIP file for CLICS submit
Move getIFiles() to EventFeedUtilities from RemoteContestAPIAdapter.  This class converts a byte stream to a zip file and extracts the contents to a List of IFile objects.
Add additional getIFiles(String) method to EventFeedUtilities to accept a BASE64 encoded string consisting of a ZIP file, then decode that and call the getIFiles(bte []) method.
Change the CLICS 2023-04 submission service to accept the previous format (a list of BASE64 encoded files) for backward compatibility, and, also accept the correct format (BAS64 encoded ZIP file).  The distinction is made by looking at the mime type of the FILE objects passed.
Update pc2submit command line utility to put the source files into a zip file, then encode that as base64 and send it off.  [This fix for pc2submit was provided by Nicky Gerritsen from the DOMjudge team.]