i_1179 CI: Add better logging for setting the contest thaw time
Setting the contest thaw time using the Contest Service PATCH facility was a bit of a black box and it was hard to determine what happened.  Now, we provide better logging.  In addition, we call the proper method to set the thaw time.
CLICSContestInfo:  be sure to include the configured thaw time value in the contest information notification message and the contests endpoint.