i_1200 Remove JudgementUtilities.getExecuteDirectoryName
This method has been removed since it is very specific to the Executable class and all references to the method in the code base should use the one in Executable since it takes into account the execute directory template and the suffix.  The JudgementUtilities version was not kept up-to-date and requires a great deal more information than ClientId currently.  This information is only availble in the Executable class, so it makes sense to use that one.